+++ engine/rating_validation/__init__.py
@@ -1,0 +1,46 @@
+from engine.rating_validation.dataset import RatingValidationDataset
+from engine.rating_validation.exceptions import (
+    DuplicateFixtureError,
+    FixtureLoadError,
+    InvalidFixtureError,
+    RatingValidationError,
+)
+from engine.rating_validation.fixture import (
+    FixtureCompleteness,
+    HattrickRatings,
+    OfficialHattrickRatings,
+    PredictedRatings,
+    RatingPredictionProvider,
+    RatingValidationFixture,
+    classify_completeness,
+)
+from engine.rating_validation.loader import load_rating_validation_dataset
+from engine.rating_validation.metrics import ErrorMetrics, calculate_error_metrics
+from engine.rating_validation.report import (
+    FixtureValidationSummary,
+    ValidationCoverage,
+    ValidationReport,
+)
+from engine.rating_validation.validator import RatingValidator
+
+__all__ = [
+    "DuplicateFixtureError",
+    "ErrorMetrics",
+    "FixtureCompleteness",
+    "FixtureLoadError",
+    "FixtureValidationSummary",
+    "HattrickRatings",
+    "InvalidFixtureError",
+    "OfficialHattrickRatings",
+    "PredictedRatings",
+    "RatingPredictionProvider",
+    "RatingValidationDataset",
+    "RatingValidationError",
+    "RatingValidationFixture",
+    "RatingValidator",
+    "ValidationCoverage",
+    "ValidationReport",
+    "calculate_error_metrics",
+    "classify_completeness",
+    "load_rating_validation_dataset",
+]